+++ app/Http/Controllers/TeamMembersController.php
@@ -49,7 +49,7 @@ class TeamMembersController extends Controller
         $notification->image_id = $team->image_id;
         $notification->url = '/teams/'.$team->slug;
         $notification->title = $team->name;
-        $notification->message = 'You have been added to the team.';
+        $notification->message = '<span class="text-green-500">You have been added to the team.</span>';
         $notification->save();
         // Trigger the event to broadcast the new notification
         event(new NewNotificationEvent($notification));
@@ -81,14 +81,33 @@ class TeamMembersController extends Controller
     {
         $teamSlug = $request->team_slug;
         $user = User::findOrFail($request->user_id);
-
         $team = Team::findOrFail($request->team_id);
+
+        // Check if the user is a manager of the team
+        if ($team->managers()->where('user_id', $user->id)->exists()) {
+            // User is a manager, proceed with detach
+            $team->managers()->detach($user->id);
+            // Additional actions after detaching the user (if needed)
+        }
+
         DB::table('teams')->where('id', $team->id)->decrement('memberSpots', 1);
         $user->teams()->detach($team->id);
 
 //        return Inertia::render('Teams/{$id}/Edit', [
 //            'members' => $team->members,
 //        ])->with('message', $user->name . ' has been successfully removed from the team.');
+
+        // notify new team member
+        $notification = new Notification;
+        $notification->user_id = $user->id;
+        // make the image the team_poster
+        $notification->image_id = $team->image_id;
+        $notification->url = '';
+        $notification->title = $team->name;
+        $notification->message = '<span class="text-orange-500">You have been removed from the team.</span>';
+        $notification->save();
+        // Trigger the event to broadcast the new notification
+        event(new NewNotificationEvent($notification));
 
         return redirect()->route('teams.manage', $teamSlug)->with('message', $user->name . ' has been successfully removed from the team.');
 //        return redirect(route('teams.manage', [$teamSlug]))->with('message', $user->name . ' has been successfully removed from the team.');