+++ resources/js/Components/Teams/Manage/TeamShow.vue
@@ -8,10 +8,10 @@
         </td>
 
         <td class="px-6 py-4 whitespace-nowrap">
-            <Link :href="`/shows/${show.slug}/manage`" class="uppercase text-xl font-semibold text-blue-600 hover:text-blue-900 dark:text-blue-200 dark:hover:text-blue-400">{{ show.name }}</Link>
+            <Link :href="`/shows/${show.slug}/manage`" class="uppercase text-xl font-semibold text-blue-600 hover:text-blue-900">{{ show.name }}</Link>
         </td>
 
-        <td class="text-gray-600 dark:text-gray-100 px-6 py-4 text-sm w-full min-w-[16rem]">
+        <td class="px-6 py-4 text-sm w-full min-w-[16rem]">
             <button v-if="!show.notes" v-show="teamStore.noteEdit !== props.show.id" class="italic" @click="editNote">Click here to add/edit a note.</button>
             <button v-if="show.notes" v-show="teamStore.noteEdit !== props.show.id" :key="componentKey" @click="editNote">{{ show.notes }}</button>
             <div v-if="teamStore.noteEdit === props.show.id">
@@ -22,19 +22,19 @@
 
         </td>
 
-        <td class="text-left text-gray-600 dark:text-gray-100 px-6 py-4 text-sm text-right">
+        <td class="text-left text-gray-600 px-6 py-4 text-sm text-right">
             {{ show.categoryName[0] }}
         </td>
 
-        <td class="text-gray-600 dark:text-gray-100 px-6 py-4 text-sm text-right">
+        <td class="text-gray-600 px-6 py-4 text-sm text-right">
             {{ show.status }}
         </td>
 
-        <td class="text-gray-600 dark:text-gray-100 px-6 py-4 text-sm text-right">
+        <td class="text-gray-600 px-6 py-4 text-sm text-right">
             {{  show.showRunnerName }}
         </td>
 
-        <td v-if="can.manageTeam || can.editTeam" class="text-gray-600 dark:text-gray-100 px-6 py-4 text-sm text-right">
+        <td v-if="can.manageTeam || can.editTeam" class="text-gray-600 px-6 py-4 text-sm text-right">
             <Link
                 v-if="can.editTeam" :href="`/shows/${show.slug}/edit`">
                 <button